+++ adminer/include/adminer.inc.php
@@ -114,14 +114,18 @@ class Adminer {
 	}
 
 	/** Get URLs of the CSS files
-	* @return list<string>
+	* @return string[] key is URL, value is either 'light' (supports only light color scheme), 'dark' or '' (both)
 	*/
 	function css(): array {
 		$return = array();
 		foreach (array("", "-dark") as $mode) {
 			$filename = "adminer$mode.css";
 			if (file_exists($filename)) {
-				$return[] = "$filename?v=" . crc32(file_get_contents($filename));
+				$file = file_get_contents($filename);
+				$return["$filename?v=" . crc32($file)] = ($mode
+					? "dark"
+					: (preg_match('~prefers-color-scheme:\s*dark~', $file) ? '' : 'light')
+				);
 			}
 		}
 		return $return;