+++ CHANGELOG.md
@@ -5,8 +5,9 @@
 - PostgreSQL, CockroachDB: Creating partitioned tables (bug #1031)
 - PostgreSQL: Move partitioned tables from table list to parent table
 - PostgreSQL: Support calling functions returning table (bug #1040)
-- Designs: adminer.css with 'prefers-color-scheme: dark' don't disable dark mode
+- Designs: adminer.css with 'prefers-color-scheme: dark' doesn't disable dark mode
 - Plugins: Method bodyClass() to add &lt;body class>
+- Plugins: Allow setting dark mode in css()
 - Hindi translation
 
 ## Adminer 5.2.1 (released 2025-04-11)